refactor: re-structure folders and files
