+++ src/components/ErrorBanner.tsx
