+++ src/public-pages/HomePage.tsx
@@ -1,0 +1,9 @@
+import Unsplash from '../unsplash/Unsplash';
+
+export default function HomePage() {
+  return (
+    <>
+      <Unsplash />
+    </>
+  );
+}
