+++ src/routes.tsx
@@ -1,14 +1,14 @@
 import { createBrowserRouter } from 'react-router-dom';
-import Layout from './public/components/Layout';
-import ErrorPage from './public/error/ErrorPage';
-import HomePage from './public/home/HomePage';
+import Layout from './components/Layout';
+import ErrorPage from './public-pages/ErrorPage';
+import Unsplash from './unsplash/Unsplash';
 
 const router = createBrowserRouter([
   {
     path: '/',
     errorElement: <ErrorPage />,
     element: <Layout />,
-    children: [{ path: '/', element: <HomePage /> }]
+    children: [{ path: '/', element: <Unsplash /> }]
   }
 ]);
